+++ src/data/codec/serde/mod.rs
@@ -1,9 +1,9 @@
-// devela::data::serde
+// devela::data::codec::serde
 //
 //! Structured serialization/deserialization.
 //
 
-mod types;
+// mod utils;
 
 crate::items! { // structural access: _mods, _all, _always
     #[allow(unused)]
@@ -12,9 +12,9 @@ crate::items! { // structural access: _mods, _all, _always
     pub use _always::*;
 
     mod _mods { #![allow(unused)]
-        pub use super::types::*;
+        // pub use super::utils::_all::*; // WIP
     }
-    pub(super) mod _all {
+    pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]
         pub use super::_mods::*;
     }